+++ resources/views/admin/account/credentials.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * @see AccountController::actionCredentials()
+ *
+ * @var View $this
+ * @var AccountUpdateForm $form
+ */
+
+use Hirtz\Skeleton\Models\Forms\AccountUpdateForm;
+use Hirtz\Skeleton\Modules\Admin\Controllers\AccountController;
+use Hirtz\Skeleton\Modules\Admin\Widgets\Forms\AccountCredentialsActiveForm;
+use Hirtz\Skeleton\Modules\Admin\Widgets\Navs\AccountSubmenu;
+use Hirtz\Skeleton\Web\View;
+use Hirtz\Skeleton\Widgets\Forms\FormContainer;
+use Hirtz\Skeleton\Widgets\Navs\Header;
+
+echo Header::make()
+    ->title($form->user->getUsername());
+
+echo AccountSubmenu::make();
+
+echo FormContainer::make()
+    ->form(AccountCredentialsActiveForm::make()
+        ->model($form));