+++ resources/views/admin/account/update.php
@@ -13,7 +13,7 @@ use Hirtz\Skeleton\Html\A;
 use Hirtz\Skeleton\Models\Forms\AccountUpdateForm;
 use Hirtz\Skeleton\Modules\Admin\Controllers\AccountController;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Forms\AccountActiveForm;
-use Hirtz\Skeleton\Modules\Admin\Widgets\Forms\TwoFactorAuthenticatorActiveForm;
+use Hirtz\Skeleton\Modules\Admin\Widgets\Navs\AccountSubmenu;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Alert;
 use Hirtz\Skeleton\Widgets\Container;
@@ -21,10 +21,12 @@ use Hirtz\Skeleton\Widgets\Forms\DeleteActiveForm;
 use Hirtz\Skeleton\Widgets\Forms\FormContainer;
 use Hirtz\Skeleton\Widgets\Navs\Header;
 
-$this->title(Yii::t('skeleton', 'Account'));
+$this->title(Yii::t('skeleton', 'COMMON_SETTINGS'));
 
 echo Header::make()
     ->title($form->user->getUsername());
+
+echo AccountSubmenu::make();
 
 if ($form->user->isUnconfirmed()) {
     echo Container::make()
@@ -42,13 +44,6 @@ echo FormContainer::make()
     ->title($this->title)
     ->form(AccountActiveForm::make()
         ->model($form));
-
-if (Yii::$app->getUser()->enableTwoFactorAuthentication) {
-    echo FormContainer::make()
-        ->title(Yii::t('skeleton', 'Two-Factor Authentication'))
-        ->form(TwoFactorAuthenticatorActiveForm::make()
-            ->model($form->user));
-}
 
 if ($form->user->isDeletable()) {
     echo FormContainer::make()